refactor: pltype
