+++ src/ast/mod.rs
@@ -5,5 +5,6 @@ pub mod diag;
 pub mod dot;
 pub mod gcapi;
 pub mod node;
+pub mod pltype;
 pub mod range;
 pub mod tokens;
